requirements.txt
aiogram==2.25.1
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async==0.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.11.2
boto3==1.26.72
botocore==1.29.72
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.6
contourpy==1.0.7
crispy-bootstrap5==0.7
cycler==0.11.0
distlib==0.3.6
Django==4.1.7
django-crispy-forms==2.0
fake-useragent==1.1.1
fastapi==0.92.0
filelock==3.9.0
Flask==1.1.2
Flask-Cors==3.0.10
fonttools==4.39.2
frozenlist==1.3.3
gunicorn==20.0.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
imageio==2.26.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==1.1.1
matplotlib==3.7.1
mechanize==0.4.8
multidict==6.0.4
networkx==3.0
numpy==1.24.2
outcome==1.2.0
packaging==23.0
PassportEye==2.2.1
pdfminer==20191125
Pillow==9.4.0
platformdirs==3.0.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.5
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==20.1
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-ip-rotator==1.0.14
rfc3986==1.5.0
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
scrapingbee==1.1.8
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.25.0
tcp-tls-tunnel==1.0.2
threadpoolctl==3.1.0
tifffile==2023.3.15
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.19.0
watchfiles==0.18.1
webencodings==0.5.1
websockets==10.4
Werkzeug==1.0.1
wsproto==1.2.0
yarl==1.8.2
